feat: basic langchain agent with gemini chore: enable roting to other screens from login chore: disable facebook login

+++ packages/server/app/database.py
@@ -4,28 +4,31 @@ from app.config import get_config
 
 settings = get_config()
 
-DATABASE_ENGINE = settings.DATABASE_ENGINE
-DATABASE_USER = settings.DATABASE_USER
-DATABASE_PASSWORD = settings.DATABASE_PASSWORD
-DATABASE_URL = settings.DATABASE_URL
-DATABASE_PORT = settings.DATABASE_PORT
-DATABASE_NAME = settings.DATABASE_NAME
+DB_CONNECTION_STRING = settings.DB_CONNECTION_STRING
 
-db_connection = f"{DATABASE_ENGINE}://{DATABASE_USER}:{DATABASE_PASSWORD}@{DATABASE_URL}:{DATABASE_PORT}/{DATABASE_NAME}"
+db_connection = DB_CONNECTION_STRING
 engine = None
 
 if settings.AMBIENTE != "pytest":
+    if not db_connection:
+        raise ValueError("Database connection string is not set.")
     engine = create_engine(db_connection)
 
 
 def create_db_tables():
+    if engine is None:
+        raise ValueError("Database engine is not initialized.")
     SQLModel.metadata.create_all(engine)
 
 
 def delete_db_tables():
+    if engine is None:
+        raise ValueError("Database engine is not initialized.")
     SQLModel.metadata.drop_all(engine)
 
 
 def get_session():
+    if engine is None:
+        raise ValueError("Database engine is not initialized.")
     with Session(engine) as session:
         yield session
